move js file
