+++ login/index.html
@@ -61,7 +61,7 @@
     </form>
   </dialog>
   <script src="../passwordEye.js"></script>
-  <script src="../loginValidCheck.js"></script>
-  <script type="module" src="../loginManager.js"></script>
+  <script src="./loginValidCheck.js"></script>
+  <script type="module" src="./loginManager.js"></script>
 </body>
 </html>
