+++ login/loginValidCheck.js
