+++ signup/index.html
@@ -73,7 +73,7 @@
     </form>
   </dialog>
   <script src="../passwordEye.js"></script>
-  <script src="../signupValidCheck.js"></script>
-  <script type="module" src="../signupManager.js"></script>
+  <script src="./signupValidCheck.js"></script>
+  <script type="module" src="./signupManager.js"></script>
 </body>
 </html>
